requirements.txt
numpy
exif
piexif
psutil
Pillow
PySide6
pillow_heif
rawpy
exifread
